[visitors] more on type handling

+++ include/crudpp/bindigs/qt/visitors/json_reader.hpp
@@ -27,7 +27,7 @@ struct json_reader
     void operator()(r_c_name auto& f) noexcept
         requires(std::floating_point<decltype(f.value)>)
     {
-        f.value = json[f.c_name()].toFloat();
+        f.value = json[f.c_name()].toDouble();
     }
 
     void operator()(r_c_name auto& f) noexcept
